类别的删除功能，表单校验

+++ src/main/java/com/zzkj/dcsystem/controller/utils/DcGoodsQueryVo.java
@@ -13,6 +13,6 @@ import lombok.Data;
 public class DcGoodsQueryVo {
     private String goodsName;
     private String typeId;
-    private Integer startPrice;
-    private Integer endPrice;
+    private Float startPrice;
+    private Float endPrice;
 }
